+++ API/src/main/java/fr/maxlego08/essentials/api/commands/Permission.java
@@ -334,7 +334,8 @@ public enum Permission {
     ESSENTIALS_DEATH_MESSAGE_TOGGLE("Allows toggling death messages visibility"),
     ESSENTIALS_DEATH_MESSAGE_TOGGLE_OTHER("Allows toggling death messages for other players"),
     ESSENTIALS_LIST,
-    ESSENTIALS_ITEMDB;
+    ESSENTIALS_ITEMDB,
+    ESSENTIALS_TERMS_ADMIN("Manage the terms of service module");
 
     private final String description;
     private final String[] args;