+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -12,6 +12,8 @@ import fr.maxlego08.essentials.module.modules.kit.KitModule;
 import fr.maxlego08.essentials.module.modules.chat.ChatModule;
 import fr.maxlego08.essentials.module.modules.customcommands.CustomCommandModule;
 import fr.maxlego08.essentials.module.modules.customscreens.CustomScreenModule;
+import fr.maxlego08.essentials.module.modules.effects.EffectsModule;
+import fr.maxlego08.essentials.module.modules.terms.TermsModule;
 import fr.maxlego08.essentials.module.modules.HomeModule;
 import fr.maxlego08.essentials.module.modules.ItemModule;
 import fr.maxlego08.essentials.module.modules.JoinQuitModule;
@@ -85,6 +87,8 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(AutoMessageModule.class, new AutoMessageModule(this.plugin));
         this.modules.put(CustomCommandModule.class, new CustomCommandModule(this.plugin));
         this.modules.put(CustomScreenModule.class, new CustomScreenModule(this.plugin));
+        this.modules.put(EffectsModule.class, new EffectsModule(this.plugin));
+        this.modules.put(TermsModule.class, new TermsModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }