+++ src/main/java/fr/maxlego08/essentials/module/modules/scoreboard/board/ClassicBoard.java
@@ -39,4 +39,9 @@ public class ClassicBoard extends FastBoard implements PlayerBoard {
     public int getLineModifier(int line) {
         return this.linesModifier.getOrDefault(line, line);
     }
+
+    @Override
+    public boolean customScoresSupported() {
+        return CustomScoresSupport.isSupported();
+    }
 }